bot/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
asyncpg==0.29.0
attrs==23.2.0
boto3==1.35.2
certifi==2024.7.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
python-telegram-bot==21.4
sniffio==1.3.1
yarl==1.9.4
numpy==1.22
pandas==1.3.3